apps/orchestrator/requirements.txt
cel-python
openai-agents
jsonschema
openai
openai-chatkit
python-dotenv
aiokafka
starlette
httpx
asyncpg
minio
